💩 : #63 - api hook과 분리
